Standarises names using SDL1 version and pang-a-lambda
